IT_asset_manager/pyproject.toml
[tool.poetry]
name = "it-asset-manager"
version = "0.1.0"
description = ""
authors = ["srush1608 <srushti.talwekar@coditas.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
langchain = "^0.3.13"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
